+++ Modules/booking/BookingRouting.php
@@ -6,6 +6,7 @@ class BookingRouting extends Routing{
 
     public function routes($router) {
         $router->map('GET', '/user/booking/future/[i:id_space]/[i:id_resource]', 'booking/booking/future', 'booking_list_future');
+        $router->map('GET', '/booking/[i:id_space]/journal', 'booking/booking/journal', 'booking_journal');
     }